+++ .eslintrc.js
@@ -1,6 +1,9 @@
 module.exports = {
   root: true,
   extends: '@react-native-community',
+  parserOptions: {
+    requireConfigFile: false,
+  },
   rules: {
     'react/no-string-refs': 'off',
     'no-alert': 'off',